feat: New way to stop the loading task
